+++ src/Compute/Compute.Test/ScenarioTests/VirtualMachineExtensionTests.cs
@@ -165,12 +165,6 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
-        public void TestAzureDiskEncryptionEncryptionWithEncryptionIdentity()
-        {
-            TestRunner.RunTestScript("Test-AzureDiskEncryptionEncryptionWithEncryptionIdentity");
-        }
-        [Fact]
-        [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void TestAzureDiskEncryptionWithIdentityNotSetInVirtualMachine()
         {
             TestRunner.RunTestScript("Test-AzureDiskEncryptionWithIdentityNotSetInVirtualMachine");